% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board_only.R
\name{board_only}
\alias{board_only}
\title{Board only}
\usage{
board_only(x)
}
\arguments{
\item{x}{A tibble (or other structure) with column "BRDPOSITION", a query field within BoardEx's "Individual Employment".}
}
\value{
Returns a filtered version of the original tibble which only contains observations for board members. Also prints a summary of how many observations were removed.
}
\description{
Board only
}
\examples{
board_only(individual_employment)
}
