% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capitalize.R
\name{capitalize}
\alias{capitalize}
\title{Capitalize}
\usage{
capitalize(x)
}
\arguments{
\item{x}{A tibble (or other structure) with columns}
}
\value{
The original tibble where all column names have been converted to uppercase
}
\description{
Capitalize
}
\examples{
capitalize(trucost)
}
