% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceo_only.R
\name{ceo_only}
\alias{ceo_only}
\title{CEO only}
\usage{
ceo_only(x)
}
\arguments{
\item{x}{A tibble (or other structure) that contains a column "ROLENAME"}
}
\value{
The structure filtered to remove any observations in which "ROLENAME" is not some form of CEO. Also prints a summary of how many observations were removed.
}
\description{
CEO only
}
\examples{
ceo_only(individual_employment)
}
