% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_fundamentals_annual.R
\name{clean_fundamentals_annual}
\alias{clean_fundamentals_annual}
\title{Clean Fundamentals Annual}
\usage{
clean_fundamentals_annual(x)
}
\arguments{
\item{x}{Tibble (or other structure) containing data from the Fundamentals Annual dataset as offered by Compustat.}
}
\value{
Cleaned version of the original tibble with NA values, overlaps, etc. removed. Also capitalizes all column names and prints a summary of what was removed.
}
\description{
Clean Fundamentals Annual
}
