% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_individual_employment.R
\name{clean_individual_employment}
\alias{clean_individual_employment}
\title{Clean Individual Employment}
\usage{
clean_individual_employment(x)
}
\arguments{
\item{x}{A tibble (or other structure) with data from a query of the Individual Profile Employment dataset as provided by BoardEx.}
}
\value{
Cleaned version of the original tibble with NA values, overlaps, etc. removed. Also capitalizes all column names and prints a summary of what was removed.
}
\description{
Clean Individual Employment
}
\examples{
individual_empl <- individual_empl \%>\% clean_individual_employment()
}
