% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_trucost.R
\name{clean_trucost}
\alias{clean_trucost}
\title{Clean Trucost}
\usage{
clean_trucost(x)
}
\arguments{
\item{x}{A tibble (or other structure) with data from a query of the Environment dataset as provided by Trucost.}
}
\value{
Cleaned version of the original tibble with NA values, overlaps, etc. removed. Also capitalizes all column names and prints a summary of what was removed.
}
\description{
Clean Trucost
}
\examples{
cleaned_env <- env \%>\% clean_trucost()
}
