% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISINtoCUSIP.R
\name{ISINtoCUSIP}
\alias{ISINtoCUSIP}
\title{ISIN to CUSIP}
\usage{
ISINtoCUSIP(x)
}
\arguments{
\item{x}{A tibble (or other structure) containing a column named "ISIN".}
}
\value{
The same tibble, but with its ISIN values converted to CUSIP values and the column name "ISIN" changed to "CUSIP"
}
\description{
ISIN to CUSIP
}
\examples{
ISINtoCUSIP(individual_employment)
}
