% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_column.R
\name{print_column}
\alias{print_column}
\title{Print Column}
\usage{
print_column(x, column)
}
\arguments{
\item{x}{A tibble (or other structure) from which to read.}

\item{column}{The name of a column in structure x to be printed to the console.}
}
\value{
The function is void; it does not return anything, only prints to the console.
}
\description{
Print Column
}
\examples{
print_column(individual_employment, ISIN)
}
