% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_column.R
\name{write_column}
\alias{write_column}
\title{Write Column}
\usage{
write_column(x, column)
}
\arguments{
\item{x}{A tibble (or other structure) from which to read.}

\item{column}{The name of a column in structure x to be printed to a tempfile.}
}
\value{
A temp_file containing the column written as a string, delineated by spaces (such that it can be copied and pasted for use with querying in WRDS).
}
\description{
Write Column
}
\examples{
write_column(individual_employment, ISIN)
}
